[P2PW-1217] - spl token send

+++ app/src/main/java/org/p2p/wallet/swap/interactor/SwapInstructionsInteractor.kt
@@ -41,7 +41,7 @@ class SwapInstructionsInteractor(
                 val transferInstruction = SystemProgram.transfer(
                     fromPublicKey = myAccount,
                     toPublickKey = accountInstructions.account,
-                    lamports = initAmount.toLong()
+                    lamports = initAmount
                 )
 
                 accountInstructions.instructions.add(1, transferInstruction)
